Move mapper to separate file
